Updating test.
